Refactored out PyMuPDF to its own plugin package
