+++ packages/paper-qa-pymupdf/src/paperqa_pymupdf/__init__.py
@@ -1,0 +1,7 @@
+from .reader import BLOCK_TEXT_INDEX, parse_pdf_to_pages, setup_pymupdf_python_logging
+
+__all__ = [
+    "BLOCK_TEXT_INDEX",
+    "parse_pdf_to_pages",
+    "setup_pymupdf_python_logging",
+]
